refactor(#4) :: statusBarHeight hook 분리
statusBarHeight를 구하는 함수를 hook으로 분리해서 사용하도록 만들었습니다.
